api: move parse_usize_param into routes/query module
utils.rs is a generic env/path helper; returning axum::http::StatusCode
from it conflated HTTP semantics with path utilities.

Extracted parse_usize_param into a new routes/query.rs module (declared
as mod query in routes/mod.rs), which keeps HTTP-specific query parsing
co-located with the route handlers that use it. utils.rs is now free of
axum and HashMap imports.

No behavior change.
