+++ apps/api/src/utils.rs
@@ -1,5 +1,3 @@
-use axum::http::StatusCode;
-use std::collections::HashMap;
 use std::env;
 use std::path::PathBuf;
 
@@ -7,17 +5,6 @@ pub fn in_dir() -> PathBuf {
     env::var("GEWEBE_IN_DIR")
         .map(PathBuf::from)
         .unwrap_or_else(|_| PathBuf::from(".gewebe/in"))
-}
-
-pub fn parse_usize_param(
-    params: &HashMap<String, String>,
-    key: &str,
-    default: usize,
-) -> Result<usize, StatusCode> {
-    match params.get(key) {
-        Some(raw) => raw.parse().map_err(|_| StatusCode::BAD_REQUEST),
-        None => Ok(default),
-    }
 }
 
 pub fn nodes_path() -> PathBuf {
